Refactor models/ & add ginkgo/gomega tests & suite
