+++ models/logger.go
@@ -1,0 +1,9 @@
+package models
+
+import (
+	"code.cloudfoundry.org/lager"
+)
+
+var (
+	helperLogger = lager.NewLogger("helper-logger")
+)
