+++ models/mocks/mocks_suite_test.go
@@ -1,0 +1,13 @@
+package mocks_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestMocks(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Mocks Suite")
+}
